Extract methods from utils/utils object to separated files
