+++ tests/collection.js
@@ -8,7 +8,6 @@
 import testUtils from '/tests/ckeditor5/_utils/utils.js';
 import Collection from '/ckeditor5/utils/collection.js';
 import CKEditorError from '/ckeditor5/utils/ckeditorerror.js';
-import utils from '/ckeditor5/utils/utils.js';
 
 testUtils.createSinonSandbox();
 
@@ -143,12 +142,6 @@ describe( 'Collection', () => {
 			'should not override item under an existing id in case of a collision ' +
 			'between existing items and one with an automatically generated id',
 			() => {
-				let nextUid = 0;
-
-				testUtils.sinon.stub( utils, 'uid', () => {
-					return nextUid++;
-				} );
-
 				collection.add( getItem( '0' ) );
 				collection.add( getItem( '1' ) );
 				collection.add( getItem( '2' ) );
@@ -157,7 +150,7 @@ describe( 'Collection', () => {
 
 				collection.add( item );
 
-				expect( item ).to.have.property( 'id', '3' );
+				expect( item.id ).to.be.above( '3' );
 			}
 		);
